+++ foodb/test/in_memory_adapter_test.dart
@@ -1,6 +1,6 @@
 import 'dart:collection';
 
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 import 'package:foodb/foodb.dart';
 import 'package:foodb/key_value_adapter.dart';