+++ foodb/test/key_value_adapter_test.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 import 'package:foodb/key_value_adapter.dart';
 
 class TestKey extends AbstractKey<String> {