+++ foodb_test/lib/src/test/delete_test.dart
@@ -1,6 +1,6 @@
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 import 'package:foodb/foodb.dart';
-import '../foodb_test.dart';
+import 'package:foodb_test/foodb_test.dart';
 
 void main() {
   // final ctx = CouchdbTestContext();
@@ -16,10 +16,10 @@ List<Function(FoodbTestContext)> deleteTest() {
       test('delete()', () async {
         final db = await ctx.db('delete');
         await db.put(
-            doc: Doc(id: "test", rev: Rev.fromString("1-a"), model: {}),
+            doc: Doc(id: 'test', rev: Rev.fromString('1-a'), model: {}),
             newEdits: false);
-        DeleteResponse deleteResponse =
-            await db.delete(id: "test", rev: Rev.fromString('1-a'));
+        var deleteResponse =
+            await db.delete(id: 'test', rev: Rev.fromString('1-a'));
         expect(deleteResponse.ok, true);
       });
     },
@@ -27,21 +27,21 @@ List<Function(FoodbTestContext)> deleteTest() {
       test('delete leafdoc', () async {
         final db = await ctx.db('delete-leafdoc');
         await db.put(
-            doc: Doc(id: "a", model: {}, rev: Rev.fromString("1-a")),
+            doc: Doc(id: 'a', model: {}, rev: Rev.fromString('1-a')),
             newEdits: false);
         await db.put(
             doc: Doc(
-                id: "a",
+                id: 'a',
                 model: {},
-                rev: Rev.fromString("2-b"),
-                revisions: Revisions(ids: ["b", "a"], start: 2)),
+                rev: Rev.fromString('2-b'),
+                revisions: Revisions(ids: ['b', 'a'], start: 2)),
             newEdits: false);
-        DeleteResponse deleteResponse =
-            await db.delete(id: "a", rev: Rev.fromString("2-b"));
+        var deleteResponse =
+            await db.delete(id: 'a', rev: Rev.fromString('2-b'));
         expect(deleteResponse.ok, true);
 
-        Doc<Map<String, dynamic>>? doc = await db.get<Map<String, dynamic>>(
-            id: "a", fromJsonT: (value) => value);
+        var doc = await db.get<Map<String, dynamic>>(
+            id: 'a', fromJsonT: (value) => value);
         expect(doc, isNull);
       });
     },
@@ -49,40 +49,39 @@ List<Function(FoodbTestContext)> deleteTest() {
       test('delete non-leafdoc, should throw error', () async {
         final db = await ctx.db('delete-non-leafdoc');
         await db.put(
-            doc: Doc(id: "a", model: {}, rev: Rev.fromString("1-a")),
+            doc: Doc(id: 'a', model: {}, rev: Rev.fromString('1-a')),
             newEdits: false);
         await db.put(
             doc: Doc(
-                id: "a",
+                id: 'a',
                 model: {},
-                rev: Rev.fromString("2-b"),
-                revisions: Revisions(ids: ["b", "a"], start: 2)),
+                rev: Rev.fromString('2-b'),
+                revisions: Revisions(ids: ['b', 'a'], start: 2)),
             newEdits: false);
         try {
-          DeleteResponse deleteResponse =
-              await db.delete(id: "a", rev: Rev.fromString("1-a"));
+          await db.delete(id: 'a', rev: Rev.fromString('1-a'));
         } catch (e) {
-          expect(e, isInstanceOf<AdapterException>());
+          expect(e, isA<AdapterException>());
         }
       });
     },
     (FoodbTestContext ctx) {
-      test("delete doc with 2 leaf nodes", () async {
+      test('delete doc with 2 leaf nodes', () async {
         final db = await ctx.db('delete-doc-with-2-leafdoc');
         await db.put(
-            doc: Doc(id: "a", model: {}, rev: Rev.fromString("1-a")),
+            doc: Doc(id: 'a', model: {}, rev: Rev.fromString('1-a')),
             newEdits: false);
         await db.put(
             doc: Doc(
-                id: "a",
+                id: 'a',
                 model: {},
-                rev: Rev.fromString("2-b"),
-                revisions: Revisions(ids: ["b", "a"], start: 2)),
+                rev: Rev.fromString('2-b'),
+                revisions: Revisions(ids: ['b', 'a'], start: 2)),
             newEdits: false);
         await db.put(
             doc: Doc(
-                id: "a",
-                rev: Rev.fromString("2-c"),
+                id: 'a',
+                rev: Rev.fromString('2-c'),
                 revisions: Revisions(ids: ['c', 'a'], start: 2),
                 model: {}),
             newEdits: false);
@@ -92,7 +91,7 @@ List<Function(FoodbTestContext)> deleteTest() {
         expect(doc?.rev, Rev.fromString('2-c'));
         expect(doc?.conflicts?[0], Rev.fromString('2-b'));
 
-        await db.delete(id: "a", rev: Rev.fromString("2-c"));
+        await db.delete(id: 'a', rev: Rev.fromString('2-c'));
         doc = await db.get(id: 'a', fromJsonT: (json) => json, conflicts: true);
         expect(doc?.rev, Rev.fromString('2-b'));
         expect(doc?.conflicts, isNull);