+++ foodb_test/lib/src/test/replicate_test.dart
@@ -1,6 +1,6 @@
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 import 'package:foodb/foodb.dart';
-import 'package:foodb/foodb_test.dart';
+import 'package:foodb_test/foodb_test.dart';
 import 'package:foodb/replicate.dart';
 
 void main() {